Moved ServiceDNSNames func to pki module
